825: Refactor compare module by removing RuboCop disable directive and moving InvalidInput class to its own file
